homefindr-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
httpx==0.28.1
pillow==11.0.0
python-dotenv==1.0.1
redis==5.2.1
celery==5.4.0
stripe==11.3.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
email-validator==2.2.0
slowapi==0.1.9
websockets==14.1
python-socketio==5.11.4
aiofiles==24.1.0
boto3==1.35.81
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
